backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
websockets==14.1
python-dotenv==1.0.1
openai==1.59.5
anthropic==0.42.0
tavily-python==0.5.0
exa-py==1.1.0
Pillow==11.1.0
httpx==0.28.1
pydantic==2.10.6
